lab6

+++ kernel/defs.h
@@ -168,7 +168,8 @@ int             copyout(pagetable_t, uint64, char *, uint64);
 int             copyin(pagetable_t, char *, uint64, uint64);
 int             copyinstr(pagetable_t, char *, uint64, uint64);
 int             ismapped(pagetable_t, uint64);
-uint64          vmfault(pagetable_t, uint64, int);
+int vmfault(pagetable_t, uint64, int);
+
 
 // plic.c
 void            plicinit(void);
